+++ changelog.md
@@ -1,7 +1,17 @@
 # ARMORED TOOLKIT CHANGELOG
 
 
-## v0.1.2 (LATEST)
+## v0.1.3 (LATEST)
+*(unreleased)*
+
+**FEATURES**
+- Keymap Override: Transform with Gizmos.
+- Operator: Transform Toggle now goes back to the previous tool instead of defaulting to Box Select.
+
+**DEV**
+- Swapped out the keymap functions for classes, simplifying the module's register/unregister methods among other things.
+
+## v0.1.2
 *(06/Jul/21)*
 - BUGFIX: Addon was deleting default blender matcaps LMAO, but only in certain custom builds.